move types into the entry file
